Install permissions for SQL account
The current documentation states that SA in SQL server is required so that it can create a login in the SQL instance. This does not adhere to least privilege principles. Instead, this account only needs to be granted access to dbcreator and securityadmin roles. I updated the SQL permissions in the section "Install required components"